matched some animation-related functions, disassembled another datablock

+++ src/ovl0/ovl0_3.h
@@ -12,17 +12,17 @@ struct AnimationUnk6C {
 
 struct AnimStack { // TODO: Same as Unk6C?
   struct AnimStack *next; // it's a list but is it really the next one?
-  u8 unk4;
+  u8 unk4; // index
   u8 unk5;
   u8 unk6;
   u8 unk7;
   f32 unk8;
   f32 unkC;
-  u32 unk10;
-  u32 unk14;
+  f32 unk10;
+  f32 unk14;
   f32 unk18;
   f32 unk1C;
-  f32 unk20;
+  s32 unk20;
 };
 
 struct Animation {
